\name{regmed-package}
\alias{regmed-package}
\alias{regmed}
\docType{package}
\title{
\packageTitle{regmed}
}
\description{
\packageDescription{regmed}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{regmed}
\packageIndices{regmed}
}
\author{
\packageAuthor{regmed}

Maintainer: \packageMaintainer{regmed}
}
\references{
Schaid, DJ, Sinnwell JP. (2020) Penalized Models for Analysis of
Multiple Mediators. Genet Epidemiol 44:408-424.

Schaid DJ, Dikilitas O, Sinnwell JP,  Kullo I. Penalized Mediation
Models for Multivariate Data. Submitted, 2021.

}
\keyword{ package }
